extract scroll y handler
