+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useFocusTrap } from './useFocusTrap';
 export { default as useOutsideClick } from './useOutsideClick';
+export { default as useScrollY } from './useScrollY';
